+++ NEWS.md
@@ -60,6 +60,7 @@ Red Panda C++ Version 2.27
   - fix: Custom compile options not correctly parsed.
   - enhancement: "Mouse scroll direction" option in Options / Editor / General
   - change: Invert scroll direction in horizontal, like in vertical.
+  - enhancement: Show type completion info after 'const' and 'volatile'
 
 Red Panda C++ Version 2.26
   - enhancement: Code suggestion for embedded std::vectors.